Foundation 4 Polyformism decimlas adjusted

+++ final/Foundation4/Swimming.cs
@@ -1,23 +1,23 @@
 class Swimming : Activity
 {
-    public int _numberLaps;
+    public double _numberLaps;
     private double _lengthLapKm;
 
     public Swimming(
         DateTime date,
-        int timeMinutes,
-        int numberLaps,
+        double timeMinutes,
+        double numberLaps,
         string typeActivity = "Swimming"
     )
         : base(date, timeMinutes, typeActivity)
     {
         _numberLaps = numberLaps;
-        _lengthLapKm = _numberLaps * 50 / 1000;
+        _lengthLapKm = _numberLaps * 50.0 / 1000.0;
     }
 
     public override double GetDistance()
     {
-        return _numberLaps * 50 / 1000;
+        return _numberLaps * 50.0 / 1000.0;
     }
 
     public override double GetPace()
